Feat: Send `userAgent` with client identity
This patch teaches session negotiation to send a user agent of the
form "com.bloomberg.bmq(javaXX.YY.ZZ):AA.BB.CC", where "XX.YY.ZZ" is
the version of Java this SDK is running on, and "AA.BB.CC" is the SDK
version.

It additionally sets the user agent to in the `PlainConsumerIT` and
`PlainProducerIT` integration test clients to different strings.
These clients are not simple users of this SDK, and construct the
sockets they communicate on directly.  It is thus useful to
distinguish them on the broker from normal users of the SDK.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>