Fix: Avoid possible `this` escape before subclass is fully initialized
Since Java16, we get a number of warnings of the sort

    [WARNING] blazingmq-sdk-java/bmq-sdk/src/main/java/com/bloomberg/bmq/impl/infr/proto/BoolMessageProperty.java:[27,25] possible 'this' escape before subclass is fully initialized

These warnings are shown when a class that may serve as a superclass
calls an overridable method in its constructor.  If some other class
inherits from that superclass and overrides that method, the
overridden implementation will be called on an object that was not
fully initialized.

We can fix this in a few ways.  On one hand, we can avoid calling
methods in class constructors.  This works nicely in cases like
`ControlMessageChoice`, which directly forwards to an `init()` method
to null out its member variables.  Duplicating this code is not
terrible.  However, in more complicated cases, like `PutHeader` and
the like, the methods we call abstract bit manupulation that’s easy to
get wrong.  It’s best not to inline that code in the constructors.

On the other hand, we can prevent the methods we call from the class
constructor from being overriden.  In all cases we have this warning,
there is a strong case to be made that we don’t want to allow
inheritance.  The message classes need to follow the protocol schema
directly, and the protocol classes are also directly tied to the
binary protocol that the BlazingMQ broker uses.  We should not allow
users to inject their own behaviors into this parsing.

This patch fixes all “`this` escape” warnings that newer JDK versions
give by marking the classes they’re shown for as `final`.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>